CNS-376: [review] keep migrations logic in a single file
